UP-3099: Refactor db_models
