+++ genai-engine/src/repositories/configuration_repository.py
@@ -1,8 +1,9 @@
-from db_models.db_models import DatabaseApplicationConfiguration
+from sqlalchemy.orm import Session
+
+from db_models import DatabaseApplicationConfiguration
 from schemas.enums import ApplicationConfigurations
 from schemas.internal_schemas import ApplicationConfiguration
 from schemas.request_schemas import ApplicationConfigurationUpdateRequest
-from sqlalchemy.orm import Session
 
 
 class ConfigurationRepository:
